+++ CHANGELOG.md
@@ -1,2 +1,8 @@
+#1.1.0
+- Remove dry run support
+- Add support for microservice requests
+- Replace Faraday with HTTParty gem
+- Refactor error handling
+
 #1.0.0
 - Initial support for microservice registration